feat: restructure user permissions handling and update related routes and services
