+++ src/common/paths.ts
@@ -28,12 +28,6 @@ export const PathsV1 = {
     tail: PathSchema.parse("/api/v1/data/tail"),
     update: PathSchema.parse("/api/v1/data/update"),
     upload: PathSchema.parse("/api/v1/data/create"),
-    perms: {
-      read: PathSchema.parse("/api/v1/data/perms/read"),
-      add: PathSchema.parse("/api/v1/data/perms/add"),
-      update: PathSchema.parse("/api/v1/data/perms/update"),
-      delete: PathSchema.parse("/api/v1/data/perms/delete"),
-    },
   },
   docs: PathSchema.parse("/api/v1/openapi/docs"),
   queries: {
@@ -55,6 +49,12 @@ export const PathsV1 = {
   user: {
     data: {
       root: PathSchema.parse("/api/v1/user/data"),
+      perms: {
+        read: PathSchema.parse("/api/v1/user/perms/read"),
+        add: PathSchema.parse("/api/v1/user/perms/add"),
+        update: PathSchema.parse("/api/v1/user/perms/update"),
+        delete: PathSchema.parse("/api/v1/user/perms/delete"),
+      },
     },
   },
 } as const;
