+++ src/data/data.controllers.ts
@@ -13,24 +13,16 @@ import {
 } from "#/middleware/capability.middleware";
 import * as DataService from "./data.services";
 import {
-  type AddPermissionsRequest,
-  AddPermissionsRequestSchema,
   type DeleteDataRequest,
   DeleteDataRequestSchema,
-  type DeletePermissionsRequest,
-  DeletePermissionsRequestSchema,
   type FlushDataRequest,
   FlushDataRequestSchema,
   type ReadDataRequest,
   ReadDataRequestSchema,
-  type ReadPermissionsRequest,
-  ReadPermissionsRequestSchema,
   type TailDataRequest,
   TailDataRequestSchema,
   type UpdateDataRequest,
   UpdateDataRequestSchema,
-  type UpdatePermissionsRequest,
-  UpdatePermissionsRequestSchema,
   type UploadDataRequest,
   UploadDataRequestSchema,
 } from "./data.types";
@@ -218,134 +210,6 @@ export function upload(options: ControllerOptions): void {
             account._id,
           ),
         ),
-        E.map((data) =>
-          c.json({
-            data,
-          }),
-        ),
-        handleTaggedErrors(c),
-        E.runPromise,
-      );
-    },
-  );
-}
-
-export function readPermissions(options: ControllerOptions): void {
-  const { app, bindings } = options;
-  const path = PathsV1.data.perms.read;
-
-  app.post(
-    path,
-    payloadValidator(ReadPermissionsRequestSchema),
-    verifyNucAndLoadSubject(bindings),
-    enforceCapability<{ json: ReadPermissionsRequest }>({
-      path,
-      cmd: NucCmd.nil.db.data,
-      roles: [RoleSchema.enum.organization],
-      validate: (_c, _token) => true,
-    }),
-    async (c) => {
-      const account = c.get("account") as OrganizationAccountDocument;
-      const payload = c.req.valid("json");
-
-      return pipe(
-        enforceSchemaOwnership(account, payload.schema),
-        E.flatMap(() => DataService.readPermissions(c.env, payload)),
-        E.map((data) => c.json({ data })),
-        handleTaggedErrors(c),
-        E.runPromise,
-      );
-    },
-  );
-}
-
-export function addPermissions(options: ControllerOptions): void {
-  const { app, bindings } = options;
-  const path = PathsV1.data.perms.add;
-
-  app.post(
-    path,
-    payloadValidator(AddPermissionsRequestSchema),
-    verifyNucAndLoadSubject(bindings),
-    enforceCapability<{ json: AddPermissionsRequest }>({
-      path,
-      cmd: NucCmd.nil.db.data,
-      roles: [RoleSchema.enum.organization],
-      validate: (_c, _token) => true,
-    }),
-    async (c) => {
-      const account = c.get("account") as OrganizationAccountDocument;
-      const payload = c.req.valid("json");
-
-      return pipe(
-        enforceSchemaOwnership(account, payload.schema),
-        E.flatMap(() => DataService.addPermissions(c.env, payload)),
-        E.map((data) =>
-          c.json({
-            data,
-          }),
-        ),
-        handleTaggedErrors(c),
-        E.runPromise,
-      );
-    },
-  );
-}
-
-export function updatePermissions(options: ControllerOptions): void {
-  const { app, bindings } = options;
-  const path = PathsV1.data.perms.update;
-
-  app.post(
-    path,
-    payloadValidator(UpdatePermissionsRequestSchema),
-    verifyNucAndLoadSubject(bindings),
-    enforceCapability<{ json: UpdatePermissionsRequest }>({
-      path,
-      cmd: NucCmd.nil.db.data,
-      roles: [RoleSchema.enum.organization],
-      validate: (_c, _token) => true,
-    }),
-    async (c) => {
-      const account = c.get("account") as OrganizationAccountDocument;
-      const payload = c.req.valid("json");
-
-      return pipe(
-        enforceSchemaOwnership(account, payload.schema),
-        E.flatMap(() => DataService.updatePermissions(c.env, payload)),
-        E.map((data) =>
-          c.json({
-            data,
-          }),
-        ),
-        handleTaggedErrors(c),
-        E.runPromise,
-      );
-    },
-  );
-}
-
-export function deletePermissions(options: ControllerOptions): void {
-  const { app, bindings } = options;
-  const path = PathsV1.data.perms.delete;
-
-  app.post(
-    path,
-    payloadValidator(DeletePermissionsRequestSchema),
-    verifyNucAndLoadSubject(bindings),
-    enforceCapability<{ json: DeletePermissionsRequest }>({
-      path,
-      cmd: NucCmd.nil.db.data,
-      roles: [RoleSchema.enum.organization],
-      validate: (_c, _token) => true,
-    }),
-    async (c) => {
-      const account = c.get("account") as OrganizationAccountDocument;
-      const payload = c.req.valid("json");
-
-      return pipe(
-        enforceSchemaOwnership(account, payload.schema),
-        E.flatMap(() => DataService.deletePermissions(c.env, payload)),
         E.map((data) =>
           c.json({
             data,
