+++ src/data/data.router.ts
@@ -8,8 +8,4 @@ export function buildDataRouter(options: ControllerOptions): void {
   DataController.tail(options);
   DataController.update(options);
   DataController.upload(options);
-  DataController.readPermissions(options);
-  DataController.addPermissions(options);
-  DataController.updatePermissions(options);
-  DataController.deletePermissions(options);
 }
