+++ src/data/data.services.ts
@@ -11,19 +11,14 @@ import { validateData } from "#/common/validator";
 import type { AppBindings } from "#/env";
 import * as SchemasRepository from "#/schemas/schemas.repository";
 import * as UserRepository from "#/user/user.repository";
+import { Permissions } from "#/user/user.types";
 import type { DataDocument, UploadResult } from "./data.repository";
 import * as DataRepository from "./data.repository";
-import {
-  type AddPermissionsRequest,
-  type DeleteDataRequest,
-  type DeletePermissionsRequest,
-  type PartialDataDocumentDto,
-  Permissions,
-  PermissionsSchema,
-  type ReadDataRequest,
-  type ReadPermissionsRequest,
-  type UpdateDataRequest,
-  type UpdatePermissionsRequest,
+import type {
+  DeleteDataRequest,
+  PartialDataDocumentDto,
+  ReadDataRequest,
+  UpdateDataRequest,
 } from "./data.types";
 
 export function createRecords(
@@ -149,67 +144,4 @@ export function tailData(
   schema: UUID,
 ): E.Effect<DataDocument[], CollectionNotFoundError | DatabaseError, never> {
   return pipe(DataRepository.tailCollection(ctx, schema));
-}
-
-export function readPermissions(
-  ctx: AppBindings,
-  request: ReadPermissionsRequest,
-): E.Effect<
-  Permissions[],
-  CollectionNotFoundError | DatabaseError | DataValidationError
-> {
-  return DataRepository.findMany(ctx, request.schema, {
-    _id: request.documentId,
-  }).pipe(
-    E.map((documents) =>
-      documents.flatMap((document) =>
-        document._perms.map((perms) => PermissionsSchema.parse(perms)),
-      ),
-    ),
-  );
-}
-
-export function addPermissions(
-  ctx: AppBindings,
-  request: AddPermissionsRequest,
-): E.Effect<
-  UpdateResult,
-  CollectionNotFoundError | DatabaseError | DataValidationError
-> {
-  return DataRepository.addPermissions(
-    ctx,
-    request.schema,
-    request.documentId,
-    request.permissions,
-  );
-}
-
-export function updatePermissions(
-  ctx: AppBindings,
-  request: UpdatePermissionsRequest,
-): E.Effect<
-  UpdateResult,
-  CollectionNotFoundError | DatabaseError | DataValidationError
-> {
-  return DataRepository.updatePermissions(
-    ctx,
-    request.schema,
-    request.documentId,
-    request.permissions,
-  );
-}
-
-export function deletePermissions(
-  ctx: AppBindings,
-  request: DeletePermissionsRequest,
-): E.Effect<
-  UpdateResult,
-  CollectionNotFoundError | DatabaseError | DataValidationError
-> {
-  return DataRepository.deletePermissions(
-    ctx,
-    request.schema,
-    request.documentId,
-    request.did,
-  );
 }
