+++ src/user/user.repository.ts
@@ -1,5 +1,5 @@
 import { Effect as E, pipe } from "effect";
-import type { StrictFilter, UUID } from "mongodb";
+import type { StrictFilter, UpdateResult, UUID } from "mongodb";
 import {
   type CollectionNotFoundError,
   DatabaseError,
@@ -12,8 +12,9 @@ import {
   type DocumentBase,
 } from "#/common/mongo";
 import type { Did } from "#/common/types";
-import type { Permissions, PermissionsDto } from "#/data/data.types";
+import type { DataDocumentBase } from "#/data/data.repository";
 import type { AppBindings } from "#/env";
+import type { Permissions, PermissionsDto } from "#/user/user.types";
 
 export type LogOperation =
   | { op: "write"; col: UUID }
@@ -149,5 +150,88 @@ export function findById(
           )
         : E.succeed(result),
     ),
+  );
+}
+
+export function addPermissions(
+  ctx: AppBindings,
+  schema: UUID,
+  documentId: UUID,
+  permission: Permissions,
+): E.Effect<
+  UpdateResult,
+  CollectionNotFoundError | DatabaseError | DataValidationError
+> {
+  const documentFilter = {
+    _id: documentId,
+  };
+  const documentUpdate = {
+    $push: {
+      _perms: permission.toJSON(),
+    },
+  };
+  return pipe(
+    checkCollectionExists<DataDocumentBase>(ctx, "data", schema.toString()),
+    E.tryMapPromise({
+      try: (collection) =>
+        collection.updateOne(documentFilter, documentUpdate, { upsert: true }),
+      catch: (cause) =>
+        new DatabaseError({ cause, message: "upsertPermissions" }),
+    }),
+  );
+}
+
+export function updatePermissions(
+  ctx: AppBindings,
+  schema: UUID,
+  documentId: UUID,
+  permission: Permissions,
+): E.Effect<
+  UpdateResult,
+  CollectionNotFoundError | DatabaseError | DataValidationError
+> {
+  const documentFilter = {
+    _id: documentId,
+    "_perms.did": permission.did,
+  };
+  const documentUpdate = {
+    $set: {
+      "_perms.$": permission.toJSON(),
+    },
+  };
+  return pipe(
+    checkCollectionExists<DataDocumentBase>(ctx, "data", schema.toString()),
+    E.tryMapPromise({
+      try: (collection) => collection.updateOne(documentFilter, documentUpdate),
+      catch: (cause) =>
+        new DatabaseError({ cause, message: "upsertPermissions" }),
+    }),
+  );
+}
+
+export function deletePermissions(
+  ctx: AppBindings,
+  schema: UUID,
+  documentId: UUID,
+  did: Did,
+): E.Effect<
+  UpdateResult,
+  CollectionNotFoundError | DatabaseError | DataValidationError
+> {
+  const documentFilter = {
+    _id: documentId,
+  };
+  const documentUpdate = {
+    $pull: {
+      _perms: { did },
+    },
+  };
+  return pipe(
+    checkCollectionExists<DataDocumentBase>(ctx, "data", schema.toString()),
+    E.tryMapPromise({
+      try: (collection) => collection.updateOne(documentFilter, documentUpdate),
+      catch: (cause) =>
+        new DatabaseError({ cause, message: "upsertPermissions" }),
+    }),
   );
 }
